ImpresionHttp/requirements.txt
altgraph==0.17.4
amqp==5.3.1
asttokens==3.0.0
async-timeout==5.0.1
billiard==4.2.1
blinker==1.6.3
celery==5.3.1
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.15
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
greenlet==3.2.3
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kombu==5.5.4
lxml==6.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pdf2image==1.17.0
pefile==2023.2.7
Pillow==10.0.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.7
pure_eval==0.2.3
Pygments==2.19.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
pywin32==306
pywin32-ctypes==0.2.3
pyzmq==27.0.0
redis==4.6.0
reportlab==4.0.4
requests==2.31.0
six==1.17.0
SQLAlchemy==2.0.21
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.3.7
win32printing==0.1.3
zipp==3.23.0
